+++ doc/changelog.rst
@@ -8,6 +8,25 @@ PyMongo 4.19 brings a number of changes including:
 
 - Added support for Python 3.15.
 
+Bug fixes
+.........
+
+- Fixed a bug where the synchronous client could permanently deadlock under
+  gevent when a greenlet was killed while checking a connection back into
+  the pool (`PYTHON-6074`_).
+
+.. _PYTHON-6074: https://jira.mongodb.org/browse/PYTHON-6074
+
+Changes in Version 4.18.1 (2026/09/10)
+--------------------------------------
+
+Version 4.18.1 is a bug fix release.
+
+- Use an exact match for the file ID in GridFS delete methods
+  (`CVE-2026-88029`_).
+
+.. _CVE-2026-88029: https://www.cve.org/CVERecord?id=CVE-2026-88029
+
 Changes in Version 4.18.0 (2026/09/03)
 --------------------------------------